checkpoint

+++ src/Modules/Blog/Explorer.Blog.API/Dtos/BlogCommentDto.cs
@@ -10,7 +10,7 @@ namespace Explorer.Blog.API.Dtos
     public class BlogCommentDto
     {
         public int Id { get; set; }
-        public int BlogId { get; set; }
+        public long BlogId { get; set; }
         public long UserId { get; set; }
         public string? Username { get; set; }
         public string? Comment { get; set; }
